Add generic analysis and decision framework
Dagger's core scheduler is pretty smart; it considers data movement
costs and expected task wait time when deciding which processor to
schedule a task on. It's always possible to make it smarter, but that
requires making tweaks across the entire codebase to handle metrics
collection, transfer, aggregation, and storage, as well as a different
set of changes to perform the analysis and come to a final decision. If
we desire to make the scheduler (and we certainly do), we're going to
need a more scalable, declarative system for adding new smarts.

This commit implements such a system, which is composed of three main
components: metrics, analyses, and decisions. Metrics are the lowest
level component, and simply generate a single measurement during the
execution of some important operation (such as task execution or chunk
movement). Analyses are the next level up, and depend on a set of
declared metrics (collected during defined operations) as well as other
analyses, producing an analysis result. Finally, Decisions are made at
specific points in Dagger's scheduler and runtime, and take in declared
analyses and metrics to produce a final decision result. This
formulation separates logic nicely such that metrics and analyses can be
written generically and reused, while decisions are focused on just
producing a result for a very specific query.

Unlike the previous implementation of the scheduler, this framework
handles all of the required metrics collection, transfer, and analysis
and decisions computations automatically and declaratively. A
user-definable top-level decision model is recursively queried to find
all necessary analyses and metrics, which are then automatically run,
cached, transferred, and made available to analyses or decisions that
require them. Only a single method needs to be specified for each
analysis and decision (for each location at which the analysis or
decision will run) to declare the required dependencies; the framework
handles the rest.

This framework makes it trivial to improve the scheduler's intelligence
with the minimum of code changes; metrics, analyses, and decisions can
be written in a minimum of 4-5 lines of straightforward code, and can be
written in a single place, such as in an external package or a user
script. This framework also defines internal helpers which make adding
new operations or decisions easy, allowing this system to expand within
the scheduler and runtime.